requirements.txt
adb_shell==0.4.4
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.1
gTTS==2.5.4
idna==3.10
numpy==2.2.3
opencv-python==4.11.0.86
packaging==24.2
pillow==11.1.0
pyasn1==0.6.1
pycparser==2.22
pydub==0.25.1
PySide6==6.8.2.1
PySide6_Addons==6.8.2.1
PySide6_Essentials==6.8.2.1
pytesseract==0.3.13
requests==2.32.3
rsa==4.9
shiboken6==6.8.2.1
sounddevice==0.5.1
soundfile==0.13.1
urllib3==2.3.0
